# perf(serde_v8): avoid SerializablePkg allocs