For common return types such as String/ZeroCopyBuf/ByteString

This should impact many ops, but here's the delta on b64 from `deno_common.js`:

Before:
```
b64_rt_long:         	n = 100, dt = 0.656s, r = 152/s, t = 6560000ns/op
b64_rt_short:        	n = 1000000, dt = 0.524s, r = 1908397/s, t = 524ns/op
```

After:
```
b64_rt_long:         	n = 100, dt = 0.632s, r = 158/s, t = 6320000ns/op
b64_rt_short:        	n = 1000000, dt = 0.456s, r = 2192982/s, t = 456ns/op
```